backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
supabase==2.24.0
httpx==0.28.1
cryptography==43.0.1
bybit-p2p==1.2.0
python-dotenv==1.0.1
numpy==2.1.3
pycountry==22.3.5

